Refactor data migration to use ResouceConfig
Use resource configuration to allow fetching for base and related
resources.
