+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/register/RegisterRepository.kt
@@ -18,7 +18,6 @@ package org.smartregister.fhircore.engine.data.local.register
 
 import com.google.android.fhir.FhirEngine
 import com.google.android.fhir.search.Search
-import java.util.LinkedList
 import javax.inject.Inject
 import kotlinx.coroutines.withContext
 import org.hl7.fhir.r4.model.Resource
@@ -26,14 +25,12 @@ import org.smartregister.fhircore.engine.configuration.ConfigType
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.configuration.app.ConfigService
 import org.smartregister.fhircore.engine.configuration.profile.ProfileConfiguration
-import org.smartregister.fhircore.engine.configuration.register.ActiveResourceFilterConfig
 import org.smartregister.fhircore.engine.configuration.register.RegisterConfiguration
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.domain.model.ActionParameter
 import org.smartregister.fhircore.engine.domain.model.ActionParameterType
 import org.smartregister.fhircore.engine.domain.model.FhirResourceConfig
 import org.smartregister.fhircore.engine.domain.model.RepositoryResourceData
-import org.smartregister.fhircore.engine.domain.model.RuleConfig
 import org.smartregister.fhircore.engine.domain.repository.Repository
 import org.smartregister.fhircore.engine.rulesengine.ConfigRulesExecutor
 import org.smartregister.fhircore.engine.util.DispatcherProvider
@@ -79,82 +76,6 @@ constructor(
       pageSize = registerConfiguration.pageSize,
       configRules = registerConfiguration.configRules,
     )
-  }
-
-  private suspend fun searchResourcesRecursively(
-    filterActiveResources: List<ActiveResourceFilterConfig>?,
-    fhirResourceConfig: FhirResourceConfig,
-    secondaryResourceConfigs: List<FhirResourceConfig>?,
-    currentPage: Int? = null,
-    pageSize: Int? = null,
-    configRules: List<RuleConfig>?,
-  ): List<RepositoryResourceData> {
-    val baseResourceConfig = fhirResourceConfig.baseResource
-    val relatedResourcesConfig = fhirResourceConfig.relatedResources
-    val configComputedRuleValues = configRules.configRulesComputedValues()
-    val search =
-      Search(type = baseResourceConfig.resource).apply {
-        applyConfiguredSortAndFilters(
-          resourceConfig = baseResourceConfig,
-          filterActiveResources = filterActiveResources,
-          sortData = true,
-          configComputedRuleValues = configComputedRuleValues,
-        )
-        if (currentPage != null && pageSize != null) {
-          count = pageSize
-          from = currentPage * pageSize
-        }
-      }
-
-    val baseFhirResources =
-      kotlin
-        .runCatching {
-          withContext(dispatcherProvider.io()) { fhirEngine.search<Resource>(search) }
-        }
-        .onFailure { Timber.e(it, "Error retrieving resources. Empty list returned by default") }
-        .getOrDefault(emptyList())
-
-    return baseFhirResources.map { searchResult ->
-      val retrievedRelatedResources =
-        withContext(dispatcherProvider.io()) {
-          retrieveRelatedResources(
-            resources = listOf(searchResult.resource),
-            relatedResourcesConfigs = relatedResourcesConfig,
-            relatedResourceWrapper = RelatedResourceWrapper(),
-            configComputedRuleValues = configComputedRuleValues,
-          )
-        }
-      RepositoryResourceData(
-        resourceRulesEngineFactId = baseResourceConfig.id ?: baseResourceConfig.resource.name,
-        resource = searchResult.resource,
-        relatedResourcesMap = retrievedRelatedResources.relatedResourceMap,
-        relatedResourcesCountMap = retrievedRelatedResources.relatedResourceCountMap,
-        secondaryRepositoryResourceData =
-          withContext(dispatcherProvider.io()) {
-            secondaryResourceConfigs.retrieveSecondaryRepositoryResourceData(
-              filterActiveResources,
-            )
-          },
-      )
-    }
-  }
-
-  /** This function fetches other resources that are not linked to the base/primary resource. */
-  private suspend fun List<FhirResourceConfig>?.retrieveSecondaryRepositoryResourceData(
-    filterActiveResources: List<ActiveResourceFilterConfig>?,
-  ): LinkedList<RepositoryResourceData> {
-    val secondaryRepositoryResourceDataLinkedList = LinkedList<RepositoryResourceData>()
-    this?.forEach {
-      secondaryRepositoryResourceDataLinkedList.addAll(
-        searchResourcesRecursively(
-          fhirResourceConfig = it,
-          filterActiveResources = filterActiveResources,
-          secondaryResourceConfigs = null,
-          configRules = null,
-        ),
-      )
-    }
-    return secondaryRepositoryResourceDataLinkedList
   }
 
   /** Count register data for the provided [registerId]. Use the configured base resource filters */
@@ -245,12 +166,6 @@ constructor(
     paramsMap: Map<String, String>?,
   ): RegisterConfiguration =
     configurationRegistry.retrieveConfiguration(ConfigType.Register, registerId, paramsMap)
-
-  private fun List<RuleConfig>?.configRulesComputedValues(): Map<String, Any> {
-    if (this == null) return emptyMap()
-    val configRules = configRulesExecutor.generateRules(this)
-    return configRulesExecutor.fireRules(configRules)
-  }
 
   companion object {
     const val ACTIVE = "active"
